ocr/requirements.txt
fastapi
uvicorn
opencv-python
pytesseract
numpy
pillow
python-multipart